+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.25;
-const unsigned char AutoSDKVersionString[] = "1.25.0";
+double AutoSDKVersionNumber = 1.26;
+const unsigned char AutoSDKVersionString[] = "1.26.0";